Rename Owin Instrumentation
