+++ examples/owin/README.md
@@ -1,7 +1,7 @@
 # OWIN Instrumentation for OpenTelemetry .NET - Example
 
 An example application that shows how to use
-`OpenTelemetry.Contrib.Instrumentation.Owin` to capture telemetry from a
+`OpenTelemetry.Instrumentation.Owin` to capture telemetry from a
 self-hosted WebAPI .NET Framework service.
 
 Span names are set to the route template resolved by WebAPI.
