--- src/OpenTelemetry.Contrib.Instrumentation.Owin/CHANGELOG.md
@@ -1,5 +1,0 @@
-# Changelog
-
-## Unreleased
-
-* Initial release
