+++ test/OpenTelemetry.Instrumentation.Owin.Tests/Controllers/TestController.cs
@@ -1,4 +1,4 @@
-﻿// <copyright file="OwinEnrichEventType.cs" company="OpenTelemetry Authors">
+﻿// <copyright file="TestController.cs" company="OpenTelemetry Authors">
 // Copyright The OpenTelemetry Authors
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
@@ -14,23 +14,17 @@
 // limitations under the License.
 // </copyright>
 
-using System.Diagnostics;
+using System;
+using System.Web.Http;
 
-namespace OpenTelemetry.Contrib.Instrumentation.Owin
+namespace OpenTelemetry.Instrumentation.Owin.Tests.Controllers
 {
-    /// <summary>
-    /// Describes the possible events fired when enriching an <see cref="Activity"/>.
-    /// </summary>
-    public enum OwinEnrichEventType
+    public class TestController : ApiController
     {
-        /// <summary>
-        /// Begin request.
-        /// </summary>
-        BeginRequest,
-
-        /// <summary>
-        /// End request.
-        /// </summary>
-        EndRequest,
+        // GET api/test/{id}
+        public string Get(string id = null)
+        {
+            return $"id:{id}";
+        }
     }
 }
